requirements/production.txt
Django==1.8
celery
redis
dj-database-url
gunicorn
honcho
psycopg2
python-social-auth
djangorestframework
djangorestframework-jwt
hashids
-e git+https://github.com/jschneier/django-storages-redux.git@47f373ef6f18e467844aa67b06cb4736d9cad958#egg=django_storages_redux-master
boto
django-imagekit
pillow
django-grappelli
flower
django-secure
django-redis
django-cachalot
django-admin-honeypot
django-notifications-hq
raven
django-summernote
django-pipeline
django-robots
